[Issue 1019] Rename `analytics.etl` sub package (#1021)
* refactor: Renamed analytics.etl to analytics.integrations
This new name more accurately reflects the purpose of the sub-package
